Refactor projects route handlers. Fix getAllProjects (fetching user projects through Model)
